64bit: * co_div64: Div whithout splitting into 32 bit * Type uint64_t for registers * '#pragma once' for faster builds. Adjust #ifndef/#define in headers
git-svn-id: svn://svn.code.sf.net/p/colinux/code/branches/devel-64bit@1498 688f6386-cf32-0410-9ead-a596f0484e40

+++ src/colinux/arch/x86_64/antinx.c
@@ -24,7 +24,7 @@
 /* 3.12 TRANSLATION LOOKASIDE BUFFERS (TLBS) */
 #define flush_tlb()							\
 	do {								\
-		unsigned long long tmpreg;				\
+		uint64_t tmpreg;					\
 									\
 		__asm__ __volatile__(					\
 			"mov %%cr3, %0;              \n"		\
