+++ src/colinux/arch/x86_64/cpuid.c
@@ -50,7 +50,7 @@ bool_t co_i386_has_cpuid()
 	return has_cpuid;
 }
 
-void co_i386_get_cpuid(unsigned long op, cpuid_t *cpuid)
+void co_i386_get_cpuid(uint32_t op, cpuid_t *cpuid)
 {
 	asm("cpuid"
 	    : "=a" (cpuid->eax),
@@ -60,10 +60,10 @@ void co_i386_get_cpuid(unsigned long op, cpuid_t *cpuid)
 	    : "0" (op));
 }
 
-co_rc_t co_i386_get_cpuid_capabilities(unsigned long *caps)
+co_rc_t co_i386_get_cpuid_capabilities(uint32_t *caps)
 {
 	cpuid_t cpuid;
-	unsigned long highest_op;
+	uint32_t highest_op;
 
 	co_i386_get_cpuid(0, &cpuid);
 	highest_op = cpuid.highest_op;
